@@ -17,7 +17,7 @@ func handleAddSSH(msg messages.IPCMessage, cfg *config.Config, vault *vault.Vaul
 	req := messages.ParsePayload(msg).(messages.CreateSSHKeyRequest)
 
 	cipher, publicKey := ssh.NewSSHKeyCipher(req.Name, vault.Keyring)
-	response, err = messages.IPCMessageFromPayload(messages.ActionResponse{
+	_, err = messages.IPCMessageFromPayload(messages.ActionResponse{
 		Success: true,
 	})
 	if err != nil {
@@ -25,10 +25,13 @@ func handleAddSSH(msg messages.IPCMessage, cfg *config.Config, vault *vault.Vaul
 	}
 
 	token, err := cfg.GetToken()
+	if err != nil {
+		actionsLog.Warn(err.Error())
+	}
 	ctx := context.WithValue(context.TODO(), bitwarden.AuthToken{}, token.AccessToken)
-	ciph, err := bitwarden.PostCipher(ctx, cipher, cfg)
+	postedCipher, err := bitwarden.PostCipher(ctx, cipher, cfg)
 	if err == nil {
-		vault.AddOrUpdateSecureNote(ciph)
+		vault.AddOrUpdateSecureNote(postedCipher)
 	} else {
 		actionsLog.Warn("Error posting ssh key cipher: " + err.Error())
 	}
