@@ -2,22 +2,21 @@ package crypto
 
 import (
 	"bytes"
-	"crypto/aes"
-	"crypto/cipher"
 	"crypto/hmac"
 	"crypto/rand"
-	cryptorand "crypto/rand"
 	"crypto/rsa"
 	"crypto/sha1"
 	"crypto/sha256"
 	"crypto/x509"
 	"errors"
 	"fmt"
-	"io"
 	"strconv"
 
 	"github.com/awnumar/memguard"
+	"github.com/quexten/goldwarden/logging"
 )
+
+var cryptoLog = logging.GetLogger("Goldwarden", "Crypto")
 
 type EncString struct {
 	Type        EncStringType
@@ -47,7 +46,7 @@ func (s *EncString) UnmarshalText(data []byte) error {
 
 	i := bytes.IndexByte(data, '.')
 	if i < 0 {
-		return errors.New("invalid cipher string format, missign type. total length: " + strconv.Itoa(len(data)))
+		return errors.New("invalid cipher string format, missing type. total length: " + strconv.Itoa(len(data)))
 	}
 
 	typStr := string(data[:i])
@@ -132,11 +131,6 @@ func DecryptWith(s EncString, key SymmetricEncryptionKey) ([]byte, error) {
 		return nil, err
 	}
 
-	block, err := aes.NewCipher(encKeyData)
-	if err != nil {
-		return nil, err
-	}
-
 	switch s.Type {
 	case AesCbc256_B64, AesCbc256_HmacSha256_B64:
 		break
@@ -158,17 +152,11 @@ func DecryptWith(s EncString, key SymmetricEncryptionKey) ([]byte, error) {
 		return nil, fmt.Errorf("decrypt: cipher of unsupported type %q", s.Type)
 	}
 
-	if len(s.IV) != block.BlockSize() {
-		return nil, fmt.Errorf("decrypt: invalid IV length, expected %d, got %d", block.BlockSize(), len(s.IV))
-	}
-
-	mode := cipher.NewCBCDecrypter(block, s.IV)
-	dst := make([]byte, len(s.CT))
-	mode.CryptBlocks(dst, s.CT)
-	dst, err = unpadPKCS7(dst, aes.BlockSize)
+	dst, err := decryptAESCBC256(s.IV, s.CT, encKeyData)
 	if err != nil {
 		return nil, err
 	}
+
 	return dst, nil
 }
 
@@ -189,19 +177,13 @@ func EncryptWith(data []byte, encType EncStringType, key SymmetricEncryptionKey)
 		return s, fmt.Errorf("encrypt: unsupported cipher type %q", s.Type)
 	}
 	s.Type = encType
-	data = padPKCS7(data, aes.BlockSize)
 
-	block, err := aes.NewCipher(encKeyData)
+	iv, ciphertext, err := encryptAESCBC256(data, encKeyData)
 	if err != nil {
 		return s, err
 	}
-	s.IV = make([]byte, aes.BlockSize)
-	if _, err := io.ReadFull(cryptorand.Reader, s.IV); err != nil {
-		return s, err
-	}
-	s.CT = make([]byte, len(data))
-	mode := cipher.NewCBCEncrypter(block, s.IV)
-	mode.CryptBlocks(s.CT, data)
+	s.CT = ciphertext
+	s.IV = iv
 
 	if encType == AesCbc256_HmacSha256_B64 {
 		if len(macKeyData) == 0 {
