@@ -16,16 +16,12 @@ import (
 
 var log = logging.GetLogger("Goldwarden", "Bitwarden API")
 
-const path = "/.cache/goldwarden-vault.json"
-
 func Sync(ctx context.Context, config *config.Config) (models.SyncData, error) {
 	var sync models.SyncData
 	if err := authenticatedHTTPGet(ctx, config.ConfigFile.ApiUrl+"/sync", &sync); err != nil {
 		return models.SyncData{}, fmt.Errorf("could not sync: %v", err)
 	}
 
-	home, _ := os.UserHomeDir()
-	WriteVault(sync, home+path)
 	return sync, nil
 }
 
@@ -34,15 +30,7 @@ func DoFullSync(ctx context.Context, vault *vault.Vault, config *config.Config, 
 	sync, err := Sync(ctx, config)
 	if err != nil {
 		log.Error("Could not sync: %v", err)
-		if allowCache {
-			home, _ := os.UserHomeDir()
-			sync, err = ReadVault(home + path)
-			if err != nil {
-				return err
-			}
-		} else {
-			return err
-		}
+		return err
 	} else {
 		log.Info("Sync successful, initializing keyring and vault...")
 	}
@@ -55,7 +43,10 @@ func DoFullSync(ctx context.Context, vault *vault.Vault, config *config.Config, 
 	}
 	if userSymmetricKey != nil {
 		log.Info("Initializing keyring from user symmetric key...")
-		crypto.InitKeyringFromUserSymmetricKey(vault.Keyring, *userSymmetricKey, sync.Profile.PrivateKey, orgKeys)
+		err = crypto.InitKeyringFromUserSymmetricKey(vault.Keyring, *userSymmetricKey, sync.Profile.PrivateKey, orgKeys)
+		if err != nil {
+			return err
+		}
 	}
 
 	log.Info("Clearing vault...")
