@@ -6,15 +6,18 @@ import (
 	"time"
 
 	"github.com/godbus/dbus/v5"
+	"github.com/quexten/goldwarden/logging"
 )
 
 var closeListenerMap = make(map[uint32]func())
 var notificationID uint32 = 1000000
+var log = logging.GetLogger("Goldwarden", "Dbus")
 
-func Notify(title string, body string, actionName string, timeout time.Duration, onclose func()) error {
+func Notify(title string, body string, actionName string, timeout time.Duration, onclose func()) {
 	bus, err := dbus.SessionBus()
 	if err != nil {
-		return err
+		log.Error("could not get a dbus session: %s", err.Error())
+		return
 	}
 	obj := bus.Object("org.freedesktop.Notifications", "/org/freedesktop/Notifications")
 	actions := []string{}
@@ -26,16 +29,17 @@ func Notify(title string, body string, actionName string, timeout time.Duration,
 
 	call := obj.Call("org.freedesktop.Notifications.Notify", 0, "goldwarden", uint32(notificationID), "", title, body, actions, map[string]dbus.Variant{}, int32(60000))
 	if call.Err != nil {
-		return call.Err
+		log.Error("could not call dbus object: %s", call.Err.Error())
+		return
 	}
 	if len(call.Body) < 1 {
-		return nil
+		return
 	}
 	id := call.Body[0].(uint32)
 	closeListenerMap[id] = onclose
 
 	if timeout == 0 {
-		return nil
+		return
 	} else {
 		go func(id uint32) {
 			time.Sleep(timeout)
@@ -45,8 +49,6 @@ func Notify(title string, body string, actionName string, timeout time.Duration,
 			}
 		}(id)
 	}
-
-	return nil
 }
 
 func ListenForNotifications() error {
@@ -62,26 +64,22 @@ func ListenForNotifications() error {
 	signals := make(chan *dbus.Signal, 10)
 	bus.Signal(signals)
 	for {
-		select {
-		case message := <-signals:
-			if message.Name == "org.freedesktop.Notifications.NotificationClosed" {
-				if len(message.Body) < 1 {
-					continue
-				}
-				id, ok := message.Body[0].(uint32)
-				if !ok {
-					continue
-				}
-				if id == 0 {
-					continue
-				}
-				if closeListener, ok := closeListenerMap[id]; ok {
-					delete(closeListenerMap, id)
-					closeListener()
-				}
+		message := <-signals
+		if message.Name == "org.freedesktop.Notifications.NotificationClosed" {
+			if len(message.Body) < 1 {
+				continue
+			}
+			id, ok := message.Body[0].(uint32)
+			if !ok {
+				continue
+			}
+			if id == 0 {
+				continue
+			}
+			if closeListener, ok := closeListenerMap[id]; ok {
+				delete(closeListenerMap, id)
+				closeListener()
 			}
 		}
 	}
-
-	return nil
 }
