@@ -45,8 +45,10 @@ func main() {
 	}
 
 	if len(os.Args) > 1 && (strings.Contains(os.Args[1], "com.8bit.bitwarden.json") || strings.Contains(os.Args[1], "chrome-extension://")) {
-		browserbiometrics.Main(&runtimeConfig)
-		return
+		err = browserbiometrics.Main(&runtimeConfig)
+		if err != nil {
+			panic(err)
+		}
 	}
 
 	cmd.Execute(runtimeConfig)
